Changed save method to bad way that works
Was working on version that only saves what it needs to but decided to
save time right now and just delete progress on each on save, and save
the current progress. This allows for a class to be taken more than
once and for easy moving and changing of classes.